requirement.txt
streamlit
torch
ultralytics
numpy
opencv-python
pillow
folium
streamlit-folium
streamlit_option_menu